requirements.txt
haystack-ai
chroma-haystack
pytest
pypdf
python-dotenv
nltk
